metrics for CR
