+++ app/etl/__init__.py
@@ -1,0 +1,1 @@
+from app.etl.metrics_etl import MetricsEtl
